feat(packaging): migrate skill tools to uv tool install
Supersedes PR #168 (symlinks + git hooks). Per Igor's review:
"let's use UV for the tools so everybody doesn't have to include their
own packaging in their script files."

Per-skill packages, installed via `uv tool install` — each tool gets
an isolated venv, entry-point binaries land in ~/.local/bin.

Changes
- Three new pyproject.toml files declaring the packages:
  - chop-gen-tts            → gen-tts
  - chop-telegram-tools     → tg-doctor, tg-watchdog
  - chop-up-to-date         → up-to-date-diag, up-to-date-hook-trust
- New install-tools.py at the repo root loops uv tool install over
  each skill package. Supports --dry-run, --uninstall, --quiet.
  Uninstall keys on package name (`uv tool uninstall chop-gen-tts`),
  never sweeps ~/.local/bin directly.
- Justfile targets: install-tools, install-tools-dry-run,
  uninstall-tools. Dropped bootstrap + core.hooksPath +
  githooks/ — uv tool install handles lifecycle.
- Script logic moved under skills/<name>/<pkg>/: each old .py now
  lives as a proper package module with `main()` / `cli_main()`
  entry-point wrappers. PEP 723 shebang blocks removed from the
  package modules; requires-python bumped to >=3.13 per convention.
- Old .py paths (skills/gen-tts/generate-tts.py, skills/harden-
  telegram/tools/*.py, skills/up-to-date/*.py) kept as deprecated
  thin shims: each `sys.modules[name] = _impl` so test patches like
  `mock.patch("watchdog.subprocess.run")` still resolve. Shims use
  PEP 723 so `./path/to/script.py --flags` continues to work for
  in-flight callers (docs, skills, crons).
- README: "Install Globally" section updated, drops the SessionStart
  hook snippet. Dropped 2026-04 symlink-era prose.
- Tests (test_diagnose.py, test_hook_trust.py, test_telegram_debug.py,
  test_watchdog.py) re-pointed at the canonical package path so
  monkey-patches hit the real module reference.

Fallback behavior
- Resource-locator functions (resolve_skill_dir in gen-tts,
  _source_dir in telegram_debug) now check $GEN_TTS_SKILL_DIR /
  $TELEGRAM_SOURCE_DIR / $CHOP_CONVENTIONS_ROOT / the default
  ~/gits/chop-conventions checkout / sibling dir — layered lookup
  keeps the common case env-var-free once installed.

Verification
- ./install-tools.py --dry-run enumerates the 3 uv tool install
  commands.
- Fresh ./install-tools.py --uninstall removes all three.
- ./install-tools.py installs all three cleanly; `which gen-tts
  tg-doctor tg-watchdog up-to-date-diag up-to-date-hook-trust`
  resolves under ~/.local/bin/.
- `tg-doctor doctor` exits 0 against live system.
- Fast tests: 91/91 up-to-date, 101/101 telegram-tools pass.
  Pre-existing failures in skills/harden-telegram/server/tests/
  test_watchdog.py (6 failures) match upstream main exactly —
  not caused by this refactor.
