+++ skills/gen-tts/pyproject.toml
@@ -1,0 +1,16 @@
+[project]
+name = "chop-gen-tts"
+version = "0.1.0"
+description = "Gemini 3.1 Flash TTS CLI — single clip or parallel batch."
+requires-python = ">=3.13"
+dependencies = ["typer>=0.12"]
+
+[project.scripts]
+gen-tts = "chop_gen_tts.cli:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["chop_gen_tts"]
