+++ skills/harden-telegram/chop_telegram_tools/__init__.py
@@ -1,0 +1,6 @@
+"""chop_telegram_tools — Telegram MCP diagnostics + watchdog packaged for `uv tool install`.
+
+Console scripts (see `pyproject.toml` for the wiring):
+    tg-doctor    → chop_telegram_tools.telegram_debug:main
+    tg-watchdog  → chop_telegram_tools.watchdog:main
+"""
