+++ skills/harden-telegram/pyproject.toml
@@ -1,0 +1,17 @@
+[project]
+name = "chop-telegram-tools"
+version = "0.1.0"
+description = "Telegram MCP diagnostics (doctor) + self-healing watchdog."
+requires-python = ">=3.13"
+dependencies = ["typer>=0.12"]
+
+[project.scripts]
+tg-doctor = "chop_telegram_tools.telegram_debug:main"
+tg-watchdog = "chop_telegram_tools.watchdog:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["chop_telegram_tools"]
