+++ skills/up-to-date/chop_up_to_date/__init__.py
@@ -1,0 +1,6 @@
+"""chop_up_to_date — `/up-to-date` skill helpers packaged for `uv tool install`.
+
+Console scripts (see `pyproject.toml` for the wiring):
+    up-to-date-diag         → chop_up_to_date.diagnose:cli_main
+    up-to-date-hook-trust   → chop_up_to_date.hook_trust:cli_main
+"""
