+++ skills/up-to-date/pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "chop-up-to-date"
+version = "0.1.0"
+description = "Git/gh repo-state diagnostic + post-up-to-date hook trust store for the up-to-date skill."
+requires-python = ">=3.13"
+# stdlib-only: diagnose.py and hook_trust.py depend on nothing outside the
+# Python standard library.
+dependencies = []
+
+[project.scripts]
+up-to-date-diag = "chop_up_to_date.diagnose:cli_main"
+up-to-date-hook-trust = "chop_up_to_date.hook_trust:cli_main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["chop_up_to_date"]
